requirements.txt
streamlit
pandas
requests
openpyxl
python-docx
altair